feat: dynamically change the OIDCProvider configuration
