+++ src/app/auth/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/components/ui/LoginForm";
+import { LoginForm } from "@/components/applications/LoginForm";
 
 export default function Page() {
   return <LoginForm />;
