+++ src/components/layout/OIDCSecure.tsx
@@ -1,0 +1,28 @@
+"use client";
+import React from "react";
+import { useRouter } from "next/navigation";
+import { useOidc } from "@axa-fr/react-oidc";
+import { useOIDCContext } from "@/hooks/oidcConfiguration";
+
+interface OIDCProps {
+  children: React.ReactNode;
+}
+
+/**
+ * Check whether the user is authenticated, and redirect to the login page if not
+ * @param props - configuration of the OIDC provider
+ * @returns The children if the user is authenticated, null otherwise
+ */
+export function OIDCSecure({ children }: OIDCProps) {
+  const { configurationName } = useOIDCContext();
+  const { isAuthenticated } = useOidc(configurationName);
+  const router = useRouter();
+
+  // Redirect to login page if not authenticated
+  if (!isAuthenticated) {
+    router.push("/auth");
+    return null;
+  }
+
+  return <>{children}</>;
+}
